+++ enclone_main/src/lib.rs
@@ -1,3 +1,4 @@
 // Copyright (c) 2021 10x Genomics, Inc. All rights reserved.
 
 pub mod main_enclone;
+pub mod split_orbits;
